backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.5.2
feedparser==6.0.11
httpx==0.27.2
python-dateutil==2.9.0.post0
APScheduler==3.10.4
beautifulsoup4==4.12.2
psycopg2-binary==2.9.11
alembic==1.13.1
urllib3>=1.26.18
requests>=2.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
redis==5.0.1
pytest-asyncio==0.24.0
playwright==1.49.1
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
scikit-learn==1.6.1
joblib==1.4.2
gunicorn==23.0.0
